+++ eas-20210701/ChangeLog.md
@@ -1,3 +1,16 @@
+2025-09-10 Version: 6.8.5
+- Update API CreateServiceCronScaler: add request parameters body.ScaleJobs.$.TimeZone.
+- Update API ListResourceInstanceWorker: add request parameters Order.
+- Update API ListResourceInstanceWorker: add request parameters Ready.
+- Update API ListResourceInstanceWorker: add request parameters ServiceName.
+- Update API ListResourceInstanceWorker: add request parameters Sort.
+- Update API ListResourceInstanceWorker: add request parameters Status.
+- Update API ListServices: add request parameters TrafficState.
+- Update API ListVirtualResource: add request parameters Order.
+- Update API ListVirtualResource: add request parameters Sort.
+- Update API UpdateServiceCronScaler: add request parameters body.ScaleJobs.$.TimeZone.
+
+
 2025-07-18 Version: 6.8.4
 - Generated python 2021-07-01 for eas.